Testing full curvature sequence. Issues with reading in the images have been solved. Now the issue is getting through the filter function.
